+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+= Taggable cache 0.4.2(unreleased)
+
+- Internal refactoring
+
 = Taggable cache 0.4.1
 
 - Fixed support for redis cache store with rails 3.2